split blob plugin apart, to make coupling more obvious
